core/tawala-ui/pyproject.toml
[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"

[project]
name = "tawala-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Kevin Wasike Wakhisi", email = "kevin@treeolive.tech" }]
requires-python = ">=3.14"
dependencies = [
    "django-phonenumber-field[phonenumberslite]>=8.4.0",
    "django-sass-processor==1.4.2",
    "libsass==0.23.0",
    "tawala-api==0.1.0",
]

[dependency-groups]
dev = ["djlint>=1.36.4", "ruff>=0.15.12", "ty>=0.0.33"]

[tool.djlint]
blank_line_before_tag = "block"
blank_line_after_tag = "endblock"

[tool.ruff]
exclude = ["*/migrations/*"]
format = { preview = true }
lint = { select = ["D"], ignore = ["D203", "D213"] }

[tool.uv.sources]
tawala-api = { workspace = true }
